src/main/java/year2022/day12/Day12.java
package year2022.day12;

import utils.FileHelper;

import java.lang.invoke.MethodHandles;
import java.util.List;

public class Day12 {

    static Object solveA(List<String> values) {

        return -1;
    }
    static Object solveB(List<String> values) {

        return -1;
    }
    public static void main(String[] args){
        var day = MethodHandles.lookup().lookupClass().getSimpleName();
        var inputs = new FileHelper().readFile("2022/"+day+".txt");

        var t0 = System.currentTimeMillis();
        var ansA = solveA(inputs);
        var t1 = System.currentTimeMillis();
        var ansB = solveB(inputs);
        var timePart1 = t1-t0;
        var timePart2 = System.currentTimeMillis()-t1;

        System.out.println(day + "A: ("+timePart1+" ms)\t"+ansA); //
        System.out.println(day + "B: ("+timePart2+" ms)\t"+ansB); //
    }
}
